Rebase branch from v3.6.1 from the latest. Prepare to merge back to
mswinport.

Conflicts:
	Makefile
	build_tools/build_detect_platform
	db/column_family.cc
	db/compaction_picker.cc
	db/transaction_log_impl.h
	include/rocksdb/env.h
	port/atomic_pointer.h
	port/port.h
	table/plain_table_index.cc
	util/dynamic_bloom.h
	util/hash_cuckoo_rep.cc
	util/ldb_cmd.cc
	utilities/merge_operators/uint64add.cc

+++ table/block.cc
@@ -304,7 +304,8 @@ Block::Block(BlockContents&& contents)
   if (size_ < sizeof(uint32_t)) {
     size_ = 0;  // Error marker
   } else {
-    restart_offset_ = size_ - (1 + NumRestarts()) * sizeof(uint32_t);
+    restart_offset_ =
+        static_cast<uint32_t>(size_) - (1 + NumRestarts()) * sizeof(uint32_t);
     if (restart_offset_ > size_ - sizeof(uint32_t)) {
       // The size is too small for NumRestarts() and therefore
       // restart_offset_ wrapped around.
